feat: Support for labeling contract vats at runtime
Closes #4738

Zoe's `installSourceBundle`, `installBundle`, and `installBundleID` methods
now accept an optional second parameter, a string used to label the bundle
installation.  If omitted it defaults to `'unlabeledContract'`.

Zoe's `startInstance` method now accepts an optional fifth parameter, a string
used to label the contract instance.  If omitted, it defaults to the empty
string (i.e., no instance label).

These two strings are used to generate the name of the contract instance's ZCF
vat.  This name currently appears in the vat worker process command line
(visible using the `ps` shell command or any other command that interrogates the
command line of running processes.  The vat name is also included as an
attribute in various vat-related slog records.)

Whereas formerly all ZCF vats had the same name, `'zcf'`, now they will be named
`'zcf-${bundleLabel}'` or `zcf-${bundleLabel}-${instanceLabel}'`.  When this
name appears in the worker process command line it also has the vatID prepended,
e.g, `v19:zcf-myAwesomeContract-instancePrime`.

Note that the vat name is not considered authoritative in any way, as a contract
may declare any bundle or instance labels it wishes, but it is helpful in
debugging a running chain in that you can more easily determine which of 47
running ZCF vats is the one you care about looking more closely at.