adjust swingStore import expectations
`importSwingStore()` has a consistency check which, for each vat,
compares the `endPos` of the current snapshot against the `startPos`
of the current transcript span.

Now that the `save-snapshot` pseudo-entry is the last one of a
transcript, these two values are expected to differ by one.

Fixed the check, updated swingstore's test-exportImport.js to behave
more like the real swingset, and added a new swingset test to exercise
the whole run/export/import/run cycle, so we can discover this sort of
thing locally in swingset's tests. Previously this only caused a
failure of the "deployment-test", which doesn't run until all other CI
has passed and the PR is at the front of the merge queue.