fix(swingset): make slog deliveryNum match transcript position
When I added pseudo-deliveries to the transcript in #7484, I forgot to
update the slog entry `deliveryNum` counter to match. As result, the
`startVat` was deliveryNum 0 but transcript position 1 (since 0 is the
`initialize-worker`). The gap grew by 2 after every snapshot save/load
event pair.

We don't slog the pseudo-deliveries as `.type: 'deliver'` (they're
recorded other ways). While it will be weird to have gaps in the
slogfile's `deliveryNum` sequence, I think it's better to maintain
easy correlation between the slog and the transcript.

So this changes `vatKeeper.nextDeliveryNum()`. Previously, that
function maintained a separate counter (in `kvStore`) for each vat,
which was only used by slog entries. Now, it queries the
transcriptStore, with `getCurrentSpanBounds().endPos`. This is only
incremented when we add an entry to the transcript, so the tests of
`nextDeliveryNum()`'s auto-increment feature were updated too.

closes #7549