fix(SwingSet): increase default snapshotInitial from 2 to 3
The new `initialize-worker` pseudo-delivery counts against our
`snapshotInitial` threshold. So the old threshold of "2" (allowing two
real deliveries to the vat before we make the first snapshot, one for
`startVat`, and another for the usual `startZcf` that our contract
vats do) needs to become "3".

This updates test-state.js to match.

A few tests provide a specific value (usually "2"), and are sensitive
to the exact sequence of deliveries being made (including
`save-snapshot` pseudo-deliveries). These were left alone.

closes #7548