fix(vats): move vat-ibc/vat-network from config to network-proposal.js
Previously, certain configurations included vat-ibc and vat-network
code as part of bootstrap.

These two vats are not properly upgradable, and are not used in our
initial launch configuration. To minimize risk, we don't want unused
code running on-chain, however we have development and test
configurations which *do* need IBC support, so we don't want to delete
this code entirely.

This commit removes ibc/network from `config.bundles`, and instead
creates "proposals" for them (just like we do for Pegasus and the
Inter protocol). You can run `agoric run
packages/vats/scripts/init-network.js` to create both a set of bundles
for installation, and a proposal file to submit in a transaction. Once
executed, this will cause the target chain to create vat-network and
vat-ibc, and to wire them up just as if they were included at
bootstrap time.

test-boot-config.js: NON_UPGRADEABLE_VATS now asserts that
'vat-network' and 'vat-ibc' do not appear in the config. The 'pegasus'
vat was removed earlier, so I uncommented that line too.

closes #7044