fix(vats): remove mintInitialSupply from basic behaviors
basic-behaviors.js included a function named `mintInitialSupply()`,
which (if run) would create the `centralSupply` vat, which can create
an initial supply of IST tokens. We obviously don't enable this in the
production config, but several unit tests and testnet configs need it,
as those environments have no other source of tokens.

To assist with auditing, this commit removes `mintInitialSupply` from
the basic behaviors entirely, so the code won't even be present in
production.

The code was moved to `demoIssuers.js`, which is used by
`sim-behaviors.js`, which is in turn used by the sim-chain and a few
tests.

refs #6687