refactor: handle missing quotes, which cause all kinds of trouble
Catch errors while calling startRound. Errors there could cause us to
miss the end of finalizing an auction round, so we would never update
nextAuctionSchedule.

If things go sideways we may end up without a next round scheduled.
If that happens, we watch for the auction params to change, and use
that to reset.