fix(auction)!: schedule capture of auction price at auction start
closes: #7898
fixes: #7794

Having renamed lockPrice to capturePrice, it becomes clear that the
time to do it is at the start of the auction. There's no need for a
separate priceLock waker.

nominalStartTime becomes more common and gets a helper.

scheduleDescendingSteps (recently renamed) gets called in a more
obvious place.

The tests get more labels showing where they are in the auction. A few
checks get updated