test: speed up these tests by at least 10 minutes
While making the regression test work, I noticed that the auctions
were configured to run on a  weekly schedule, and it was taking the
timer a long time to cycle through to the next auction. I changed the
other tests to default to one hour auctions except in one case that
wanted to charge interest, and was fast anyway. The test suite went
from about 12 minutes to run to 1.5 minutes.