fix(vaults!): distribute all Collateral after liquidation
fixes: 7806

In flow 2b, the penalty was subtracted out at the beginning of
distributing liquidation proceeds, and wasn't included in the amount
sent to the Reserve.

This fix sends all the remainder to the Reserve (which will pick up
the leftovers the first time it is run), and warns if it doesn't match
the expected amount.