fix(xsnap): use newer xsnap-pub, with requirement for __has_builtin
This switches to a version of `xsnap-pub` that adds a compile-time
assertion of the availability of `__has_builtin()`. This function was
added to gcc-10, and was missing in gcc-9. XS can produce different
heap snapshots depending upon whether this function is available or
not (breaking consensus, which includes the heap snapshot hashes). By
mandating its presence, we ensure that all instances will behave the
same way, maintaining consensus.

closes #7829