fix(vaults): address a divide-by-zero; correct penalty w/no bidders
 #7785 calculated the price before it was needed because it could have
 been useful in two branches. Unfortunately, in the third branch the
 denominator would often be zero, which makeRatio rejects.

This moves the price calculation back to where it's needed, and where
the denominator is know to be positive.

In doing so, I discovered that there was no test coverage for cases
where no collateral was sold. In those cases, the penalty charged
against vaults being reinstated was calculated incorrectly, so the
vault wouldn't have been reinstated. This corrects that.