chore!: rm getMetrics from unpublished contracts
PSM still has it because it is on-chain and there's a higher bar for breaking changes there. VaultFactory and Reserve contracts haven't yet been deployed on chain. This arguably isn't a breaking change at all, but will be to anyone using these NPM packages to get contracts they then run.