rewrite importSwingStore to preserve metadata properly
Previously, the swingstore importer would ignore "historical
metadata": records (with hashes) for transcripts and heap snapshots
that are not strictly necessary to rebuild workers. This was a
mistake: our intention was to always preserve these hashes, so that we
might safely (with integrity) repopulate the corresponding data in the
future, using artifacts from untrusted sources.

This commit rewrites the importer to record *all* metadata records in
the first pass, regardless of whether we want historical data or
not. All of these records will be stubs: missing the actual bundle or
snapshot or transcript items, as if they had been pruned. Then, in the
second pass, we populate those stubs using the matching artifats (or
ignore the historical ones, as configured by the `includeHistorical`
option). A final `assertComplete` pass insists that all the
important (non-historical) records are fully populated.

New tests were added to assert that metadata is preserved regardless
of import mode, and that the `assertComplete` pass really catches
everything.

A new `docs/swingstore.md` was added to describe the data model,
including what it means for records to be pruned.

fixes #8025