swingstore: add repairMetadata()
Since bug #8025 caused swingstore imports to omit historical metadata,
there may be deployed instances that are missing data. To fix this,
the new `repairMetadata()` function will take a pre-existing
swingstore and an (intact) exporter, and will re-create any metadata
records that don't already exist.

The function will reject any records that do exist but whose contents
do not match the incoming data. It will also completely ignore any
kvStore records (which are not really metadata). It performs a commit
at the end, and does not return a SwingStore instance.

refs #8025