feat(swingstore): add repairMetadata()
Since bug #8025 caused swingstore imports to omit historical metadata,
there may be deployed instances that are missing data. To fix this,
the new `hostStorage.repairMetadata()` method takes an (intact)
exporter, and will re-create any metadata records that don't already
exist. This function ignores artifacts entirely.

The method will throw if the metadata stream has records that already
exist in the DB but whose contents do not match. It also throws if the
stream creates a new bundle record, since without artifacts that new
record cannot possibly be populated, and the data model requires all
bundles be present. The method ignores any kvStore records (which are
not really metadata).

Afterwards, the caller is responsible for calling
`hostStorage.commit()` when they are ready.

refs #8025