fix: rewrite importSwingStore to preserve metadata properly
Previously, the swingstore importer would ignore "historical
metadata": records (with hashes) for transcripts and heap snapshots
that are not strictly necessary to rebuild workers. This was a
mistake: our intention was to always preserve these hashes, so that we
might safely (with integrity) repopulate the corresponding data in the
future, using artifacts from untrusted sources.

This commit rewrites the importer to record *all* metadata records in
the first pass, regardless of whether we want historical data or
not. All of these records will be stubs: they contain hashes, but are
missing the actual bundle or snapshot or transcript items, as if they
had been pruned. Then, in the second pass, we populate those stubs
using the matching artifacts (or ignore the historical ones, as
configured by the `includeHistorical` option). A final
`assertComplete` pass insists that all the important (non-historical)
records are fully populated.

The exporter was updated to omit empty artifacts.

New tests were added to assert that metadata is preserved regardless
of import mode, and that the `assertComplete` pass really catches
everything. Also, we check that an import throws if given a mis-sized
artifact, like a transcript span that is missing a few items.

A new `docs/swingstore.md` was added to describe the data model,
including what it means for records to be pruned, and
`docs/data-export.md` was updated.

Note: this commit changes the schema of the `snapshots` table (to
support temporarily-unpopulated `inUse = 1` snapshot data). To be
precise, any swing-store created by this version (either via
`initSwingStore` or `importSwingStore`) will get the new schema:
pre-existing DBs opened with `openSwingStore` will continue to use the
old/strict schema. This is fine for now, but as the comments in
snapStore.js explain, we'll need to implement DB schema versioning and
upgrade (https://github.com/Agoric/agoric-sdk/issues/8089) before we
can safely change any non-`importSwingStore` code to create
unpopulated `inUse=1` records.

fixes #8025