feat: implement swingset-runner based version of the vaults benchmark
Note: This commit also includes the stripped version of the Ava-based vaults
benchmark, which is like the regular Ava-based benchmark but has had all the
timing and data collection code removed from it (making it much shorter and
easier to read) in the expectation of external rather than internal measurement.
This was an engineering stepping stone to getting the benchmark to run inside a
swingset, but is being committed to the repo because we suspect it may be a
useful reference in the future.