add packages/ghost-chain
To run, first make a copy of a mainnet swingstore DB in a new
`swingstore/` directory, then run:

 node .../packages/ghost-chain/src/run-ghost.js .../swingstore SUBCOMMAND ARGS

There are several SUBCOMMANDs, one is `terminate-vat v46`.

This will write to several files:
* `init.slog`: capture the activity of loading the kernel and
preloading vats
* `upgrade.slog`: everything that happens after the subcommand is
executed
, `cleanup.slog`: capture the subsequent series of cleanup runs
* `bridge.out`: writes to the bridge device

The tool writes normal console logs to stdout, followed by extensive
notes on kvStore changes.

The tool modifies and commits the DB in place, so it is
destructive. You can comment out that `commit()` if you don't need it,
in which case 1: the DB will be left unmodified (making it easier to
run the tool multiple times, without re-copying the original
swingstore), and 2: RAM/.wal usage will be higher (and performance
will be less realistic), as all DB changes are instead buffered.