fix: eliminate the `passableEncoding` hack
`PassableEncoding` (not to be confused with `encodePassable`) was a non-standard
serialization scheme used in some tests, which encoded remote references in-band
using Symbols with magic names rather than using the normal marshal package
machinery that puts these into the 'slots' element of the standard capdata
format.  This bypassed various message filtering and transformation logic in the
kernel and also required special methods to be present in the bootstrap vat to
translate this encoding and relay messages to their actual intended
destinations.

This has now been removed.  The relatively small number of tests which used
`passableEncoding` have been updated to use `kmarshal` instead.  Messages and
data are now encoded in a form that all the other code understands.  Test
messages are also now delivered directly to their destinations without having to
count on the existence of a relayer.

In support of this, the controller's `queueToVatRoot` method has been augmented
by the addition of a `queueToVatObject` method, allowing tests to send messages
to specific objects, targeted using remotable references of the sort returned
by `kunser`.  The test support library that a lot of the bootstrap tests use
has been updated to use this improved mechanism.

In addition, `kmarshal` itself has been upgraded using a trick that MarkM
provided for tagging promises, which allows `kmarshal` to be truly stateless.
The (former) statefulness of `kmarshal` caused problems when the module was
imported into different compartments, as each compartment ended up with its own
module instance and thus its own version of the state. This in turn caused these
compartments to have different beliefs about how particular promises were
represented, which caused various things to break.  That's all fixed now.

One wart which has NOT been taken care of in this PR, but which will be
addressed in a follow-on PR that we were already planning for, is the
duplication of `kmarshal.js` in both the SwingSet package and the liveslots
package.  The forthcoming PR will perform a bunch of file renaming and
relocation to put a bunch of support tooling, used by both benchmarks and tests,
into a package of its own, thereby eliminating a lot of weird dependencies and
files in places they don't belong.  As part of this I plan to relocate
`kmarshal` into a package of its own that can then be cleanly imported by the
kernel, liveslots, and the various tests and test support tooling.

All this is in support of issue #8327