fix(smartWallet): handle upgrade disconnects from purse notifiers
observeNotifier() would handle upgrade disconnects if the notifiers
were durable. But in @agoric/ertp, they're ephemeral, so we open-code
the loop.

When the offerWatchers detect an upgrade, they reschedule the
watchers. We needed to ensure the wallet itself doesn't do its cleanup
in that case.

The test updates purse balance across upgrade
  - provision a smartWallet for an oracle operator
  - upgrade zoe; reproduce smartWallet bug
  - check for new invitation