build: lint all types
New command, lint:all-types, that does a typecheck over the whole repo at once.
On my machine it takes 7s, versus minutes for typechecking within each package.

We want both to continue working so this leaves those alone and adds new
command that runs in the toplevel.

This isn't yet in CI.