Zoe use watchPromises for the release branch (#8710)
* feat: zcf ask zoe to repair 'adminNode.done()` watcher on restart

use watchPromise() to wait for contract finish; This repairs new
contracts, but doesn't help existing contracts.

* test: add tests for userSeat.getExitSubscriber()

* test: repair broken test of upgrade

There was a bug in the part of the test that was attempting to run a
bunch of scenarios across the upgrade. Rather than resuming each run
after the upgrade, it was running the whole scenario from scratch
after the upgrade.

* test: demonstrate Zoe's survival through vatAdmmin vat upgrade

* refactor: upgrade-zoe proposal name

it's not necessarily null. it's whatever the version is that's built.

---------

Co-authored-by: Turadg Aleahmad <turadg@agoric.com>