In building the a3p-integration image with the PFM was seeing an error like: `failed to load latest version: version of store packetfowardmiddleware mismatch root store's version; expected 1337 got 0`
Found https://github.com/cosmos/cosmos-sdk/issues/16003 which hinted "Probably forgot to configure the store upgrades in a upgrade that added this store." The upgrade guide for PFM didn't mention this but I got lucky and remembered the `Delete` block about `x/crisis`. Created an `Added` block in the same place and things got happy.