test(async-flow): demonstrate #9465 (#9474)
closes: #XXXX
refs: #9465 

## Description

This PR just captures the test from #9465 , turning it into an expected
failure. A later PR that fixes the bug would then remove the `.failing`.

### Security Considerations

The bug itself endangers integrity, so fixing it will help security.
This PR though just adds the test that demonstrates the bug.

### Scaling Considerations

none

### Documentation Considerations

none
### Testing Considerations

The point. 
### Upgrade Considerations

The bug means that an asyncFlow activation that does return early would
be committed to the done state, preventing a later upgrade from
repairing it to get beyond the problem. But this PR just adds the test
for that bug.