fix(vow): watcher args instead of context (#9556)
closes: #9555

## Description

This PR updates the internal watcher exo to handle an args list, and updates the `watch` tool to check the arity and realize if a 3rd argument has been provided or not. It does not currently accept a regular rest argument like `watchPromise` does.

### Security Considerations
None

### Scaling Considerations
None

### Documentation Considerations
Updated type definition

### Testing Considerations

I didn't find any unit test coverage for this but some integration tests have shape on the watcher checking the arity.

### Upgrade Considerations
Need to settle the internal handling of arguments before we commit to them in state.