style(swing-store): Use the "generator" convention in function names (#9650)
Ref #9423

## Description
This maintains `getArtifactNames` as a SwingStoreExporter method that synchronously either throws an error or returns an async iterator, but renames the internal helper function responsible for constructing that iterator from the overly-similar `artifactNames` to the more conventional `generateArtifactNames`. I initially wanted to reconsolidate both behaviors into `getArtifactNames`, but ultimately decided that separation better highlights the potential for a synchronous exception.

### Security Considerations
n/a

### Scaling Considerations
n/a

### Documentation Considerations
n/a

### Testing Considerations
n/a

### Upgrade Considerations
n/a