fix: adopt `VTRANSFER_IBC_EVENT` as an action-type (#9671)
closes: #9670

Ensure that the `VTRANSFER_IBC_EVENT` messages are plumbed through the `cosmic-swingset` layer.

Prevents chain halt when the `x/vtransfer` facilities are used.

This bug was overlooked because end-to-end testing of this facility was not completed until a late stage. Earlier unit testing used mocks of the bridge device that did not have the same problematic behaviour. More automated end-to-end IBC testing in a close-to-production setting is needed.

Fixing this `cosmic-swingset` behaviour does not require upgrade of any other components.