feat: add the ability to override param values
When upgrading a contract, some parameter values originally defined
based on terms must be overridden by values supplied via
privateArgs. This provides a path to supply those overrides when
creating the new paramManager.

The vaultFactory uses this new feature to pass values through.

The proposal to upgrade vaultFactory gets the current values as
updated by governance and passes them in when upgrading.