fix(liveslots): collection.clear(pattern) vs getSize
The `.clear()` method, on (strong) collections, will delete all entries.
But it can also be called with `keyPatt` and/or `valuePatt` argument,
and only the matching entries will be deleted.

Previously, the `|entryCount` metadata field was unconditionally set
to zero for any call to `.clear()`, even those with patterns which
caused some entries to be retained. This caused a subsequent
`collection.getSize()` to return the wrong value, yielding zero even
if the collection still had entries.

This commit fixes the logic to decrement the `|entryCount` field by
the number of entries actually deleted. As an optimization, it is
still set directly to zero if `clear()` is not limited by patterns.

fixes #10007