Improve xsnap-worker's reaction to EOF/SIGPIPE by explicit logging to stderr (#4103) (#10099)
closes: #4103 
refs: [agoric-labs/xsnap-pub PR #53](https://github.com/agoric-labs/xsnap-pub/pull/53)

## Description
This is the agoric-sdk side of the changes for xsnap-worker, containing updating the git subproject and adding new tests.

### Security Considerations
New version of xsnap-worker explicitly exits and logs clear text errors to stderr when it detects an unexpected EOF or SIGPIPE condition, which does disclose some information regarding its internal state.

### Scaling Considerations
Overall behavior is the same as xsnap-worker dies/exits just as before.

### Documentation Considerations
none for agoric-sdk.

### Testing Considerations
a new test had been added.

### Upgrade Considerations
none.