fix(swingset): change processUpgradeVat to delete clist entries
The kernel is responsible for "disconnecting" (rejecting) promises
decided by a vat being upgraded, because Promise settlement
functions (resolve/reject) are ephemeral, so they are lost during the
upgrade, so the new incarnation will have no way to ever settle them.

When the vat resolves a promise with syscall.resolve(), that vat's
c-list entry is removed by the syscall handler. But when the kernel
did the rejection, the c-list entry was left in place, causing the
promise to be pinned forever.

This changes the kernel to remove the c-list entry as well, except in
the case that the vat is also subscribed to its own promise (which
happens when a PromiseWatcher is used on a locally-decided promise:
unusual but we needed to tolerate that case). Self-subscribed promises
are left in the c-list, because the subsequent dispatch.notify() will
finish the job and remove the entry.

This adds unit tests for the new behavior (including the
self-subscription case). It does not yet add remediation
code. Therefore it merely:

refs #9039

rather than fixing it entirely.