change kernelKeeper.enumeratePromisesByDecider() API to be more efficient
Previously, this function identified a set of promises (unresolved
promises decided by a particular vat) and yielded an iterator of their
kpids. To perform this filtering, it extracts their promise records
first, but throws that data out and only uses the kpids.

Some clients will need the full promise record, so this can be more
efficient if we return both the record and the kpid, and let the
caller discard the parts they don't need.