chore: more efficient kernelKeeper.enumeratePromisesByDecider() API
Previously, this function identified a set of promises (unresolved
promises decided by a particular vat) and yielded an iterator of their
kpids. To perform this filtering, it extracts their promise records
along the way, but then throws out that data and only uses the kpids.

Some clients will need the full promise record, so this can be more
efficient if we return both the record and the kpid, and let the
caller discard the parts they don't need.

This also improves one client: the code that iterates over dead
promises when terminating a vat. Rather than build a list of kpids in
RAM, we can simply resolve them one at a time, since we are no longer
in danger of the kvStore records being deleted before we resolve
them (thanks to slow/deferred deletion).