refactor: set roundId to 1 & write to vstorage at start
Use a minimum start time and roundId of 1 because zero is treated as
an outlier in the code

There are places in the code and tests that compare roundId and round
start time to 0 and 1, so make it easier by starting from 1.

Refactored some tests, mostly by adding a round before the previous tests
to let state get past zero and one.

refactor: set roundId to 1 & write to vstorage at start

use a minimum start time and roundId of 1 because zero is treated as
an outlier in the code

There are places in the code and tests that compare roundId and round
start time to 0 and 1, so make it easier by starting from 1.

Refactored some tests, mostly by adding a round before the previous tests
to let state get past zero and one.