fix(ERTP,vats): fix 9407 AmountPatternShape (#9863)
Staged on #9410 

closes: #9407 
refs: #9410 

## Description

Originally #9410 was supposed to fix #9407 , by relaxing an incorrect rejection into a correct acceptance. But we are not yet ready to deploy such a fix on-chain, due to upgrade constraints, so fixing it in agoric-sdk master would be misleading. It might lead people to write other code that depends on this bug being fixed when tested against agoric-sdk master, only to have that code fail later on-chain when invoking the non-upgraded buggy code.

So we unfixed #9410 to restore the buggy behavior, but with a better error message, and refactored to make fixing it again easier. This PR is that fix. It will remain draft until we are confident we can upgrade the relevant on-chain vat (the vbank) in a timely manner. Nevertheless, though we'll keep it in draft until then as a safety measure, this PR is reviewable. Reviewers, please review it.

### Security Considerations
Fixing a bug is generally good for security. Otherwise, none.

### Scaling Considerations
none

### Documentation Considerations
Would remove the need to document the #9407 buggy behavior

### Testing Considerations
Restores the `test.failing` from #9410 to a `test`.

### Upgrade Considerations
as above.