fix(ertp): remove unneeded ertp type imports (#10467)
closes: #XXXX
refs: #XXXX

## Description

For some reason, vscode does successfully highlight unused jsdoc type imports

![image](https://github.com/user-attachments/assets/8ba7670d-cde2-4458-b5dc-501b4b315414)

but `yarn lint` does not

![image](https://github.com/user-attachments/assets/1f4769f4-2f85-41f2-886b-9a0e6c5a0078)

This PR started as a drive-by extracted from https://github.com/Agoric/agoric-sdk/pull/10456 , where I noticed that ERTP in particular had some of these. Since #10456 experiments with a variation on ERTP, simplifying it first in harmless ways helps a bit.

### Security Considerations

fewer distractions helps

### Scaling Considerations

none

### Documentation Considerations

none

### Testing Considerations

none

### Upgrade Considerations

In the agoric-sdk repo there are no remaining type imports of the legacy type reexport of `Baggage` from ertp. But if there are such imports in other repos, they will need to be fixed. But this is a static-only issue, not a runtime issue, so it's fine for those repos to be fixed only once they depend on the new `@agoric/ertp` without that legacy reexport.