chore: Better accommodate non-GNU sed (#10482)
## Description
Always use `sed -i …` vs. `sed -i ''` as appropriate, with a `sedi` function or alias where possible.

### Security Considerations
n/a

### Scaling Considerations
n/a

### Documentation Considerations
n/a

### Testing Considerations
I haven't found a practical way to test this, so it might go stale—but if so, it's only in development/build tooling so it can be fixed again if that happens.

### Upgrade Considerations
n/a