feat(fast-usdc): write chain policies to vstorage (#10532)
closes https://github.com/Agoric/agoric-sdk/issues/10512

## Description

Adds the shape/type for the chain policies, and makes the fast-usdc.start script write them to vstorage

### Security Considerations

Only the start script and the contract have permission to write to the vstorage node

### Scaling Considerations

This is a single invocation that runs once per core-eval, so shouldn't impact performance over time.

### Documentation Considerations

There's an example policy in the test snapshot

### Testing Considerations

Wrote a bootstrap test that reads vstorage to verify the chain policies are published.

### Upgrade Considerations

The core eval would have to write to vstorage again to update the policies