test: After updating vaultFactory params, verify the change (#10803)
## Description

We noticed that after changing params for the VaultFactory, the test verified that the vote had completed, but not that the param change was effective. For a while we suspected that the change didn't take, but a simple enhancement to the test showed that wasn't the case.

### Security Considerations

None

### Scaling Considerations

None

### Documentation Considerations

None

### Testing Considerations

Tests should verify the actual changes they make.

### Upgrade Considerations

This was one red herring we chased while validating upgrade. A clearer test would have made that chase shorter.