fix(ci): Race condition in wait-for-follower script (#11159)
closes: https://github.com/Agoric/product-tasks/issues/264
refs: #XXXX

## Description
Due to a possible race condition in the [wait-for-follower.mjs](https://github.com/Agoric/agoric-sdk/blob/master/a3p-integration/proposals/z%3Aacceptance/wait-for-follower.mjs) script, the CI job can hang. The script first checks for the current contents of the file and then starts a watcher on the file. If the intended message was written to the file in between, the script never exits and the CI job hangs.
This PR adds a wrapper on the file watcher and reports a fake event first to check the existing file contents before starting the actual file watcher

### Security Considerations
None

### Scaling Considerations
None

### Documentation Considerations
None

### Testing Considerations
CI build should pass

### Upgrade Considerations
None