fix: `Observed` is not a `PendingTxStatus`
`Observed` is a state only controlled by the `TransactionFeed`. It's published to vstorage and is a valid `TxStatus`,
but it is not part of the `StatusManager`, `Advancer`, or `Settler` state machines.

`PendingTx`s are initialized as either `Advancing` or `AdvanceSkipped`. An `Advancing` can be transitioned to `Advanced`
or `AdvancedFailed`. From here, once a matching mint arrives, they are dequeued from the store (no longer pending) and
are ultimately transitioned to `Disbursed` or `Forwarded`