feat(base-zone): add durable revocation/cancellation tools
`prepareInternalRevocableMakerKit` extracts the exo class
kit that is wrapped by the `prepareRevocableMakerKit` facade.
Although the facade is easier to use, it is not an exo, so
its facets cannot be stored to survive upgrade.

`prepareCancelKit` has `guard` and `canceller` facets to be
used by arbitrary functions.  The guard reveals whether or
not the canceller has been called.