fix(cosmos): snapshot init error is not fatal (#11903)
refs: https://github.com/Agoric/agoric-sdk/pull/11378/files#r2332324935

## Description
When updating cosmos app to 0.50, the signature of `Commit` changed to add an error return value instead of relying on explicit panic. Previously an error in snapshot initiation was just reported but didn't panic. The equivalent would have been to not return the error in that case.

### Security Considerations
None

### Scaling Considerations
None

### Documentation Considerations
Added a comment to clarify the error is just logged

### Testing Considerations
In a perfect world we'd have a test of this, but app.go is not exactly unit testable and integration testing this behavior is pretty onerous.

### Upgrade Considerations
Requires chain software upgrade. This should be cherry-picked in u22.