ci(Mergify): don't label on queuing (#12023)
Disable the new `queued_label` as that negatively interacts with our
"merge-strategy" check. When mergify embarks a PR, it would label the
PR, which would temporarily invalidate the conditions that got the PR
embarked in the first place, causing a race on whether the check
finishes first, or whether mergify checks the conditions again first.