ci: generic bundle builder workflow (#12144)
refs: #12136
closes: #12142

## Description
Adds a generic workflow to run `build:bundles` in a package. Takes a package name and branch as input and builds bundles against it. Requires the package to have `yarn build` and `yarn build:bundle` which does the needful. Moving forward, it is expected that *deploy packages will have these yarn scripts that generate the necessary bundles. The bundles generated in the `dist` folder of the provided package are then uploaded as GH artifact for re-use. 

### Security Considerations

None

### Scaling Considerations

None

### Documentation Considerations

None, internal use only.

### Testing Considerations

[Tested](https://github.com/Agoric/agoric-sdk/actions/runs/18685598045/job/53277244205?pr=12144) on `master` via some slight modifications to the workflow to allow running it on PRs. 

### Upgrade Considerations

None.