feat!(smart-wallet): do not publish invitation details handles (#12167)
refs: #10744
closes: #12160

## Description
While https://github.com/Agoric/agoric-sdk/pull/12085 added a local caching marshaller to avoid unnecessary round trip to the board, the smart wallet still needs to hit the board anytime it holds and wants to serialize an invitation detail. The reason is that invitation details have a unique `handle` that comes back as a severed reference from the board (`null` slot).

Even for continuing invitations where the invitation details don't change, the caching marshaller cannot assume that because a capability wasn't in the board at one point it will not be in the future, so it takes the safe approach of never caching severed references.

This change avoids sending these handles to the board in the first place. It does so by filtering any published balance amounts that contain an invitation detail. As such it affects both the publishing to vstorage and the public topics, and in a sense aligns both type of consumers. Smart wallet public topics are unused, and off chain clients only use the `description` and `instance` fields of invitations details published to vstorage to form invitation queries (since handles are severed in the first place). As such the filtered invitation details remain a super set of the invitation query data.

An alternative to changing the shape of published data would be for the smart wallet to "inject" a severed resolution into the cache of the marshaller, but to maintain the independence of smart wallets that would require creating a different caching marshaller per wallet. That's a much more expensive change.

### Security Considerations
None

### Scaling Considerations
Reduce gc pressure created by the constant sharing and dropping of capabilities with the board.

### Documentation Considerations
None

### Testing Considerations
Manually verified in the slog created by the fast-usdc chain impact test that the smart wallet would no longer invoke `toCapData` of the board marshaller when using or holding invitations, and that the value written to vstorage lacks a severed handle reference

```json
{"type":"syscall","crankNum":5647,"vatID":"v25","deliveryNum":114,"syscallNum":0,"replay":false,"syscall":"send","ksc":["send","ko670",{"methargs":{"body":"#[\"setValue\",[\"{\\\"body\\\":\\\"#{\\\\\\\"currentAmount\\\\\\\":{\\\\\\\"brand\\\\\\\":\\\\\\\"$0.Alleged: Zoe Invitation brand\\\\\\\",\\\\\\\"value\\\\\\\":[{\\\\\\\"description\\\\\\\":\\\\\\\"oracle operator invitation\\\\\\\",\\\\\\\"installation\\\\\\\":\\\\\\\"$1.Alleged: BundleIDInstallation\\\\\\\",\\\\\\\"instance\\\\\\\":\\\\\\\"$2.Alleged: InstanceHandle\\\\\\\"}]},\\\\\\\"updated\\\\\\\":\\\\\\\"balance\\\\\\\"}\\\",\\\"slots\\\":[\\\"board0074\\\",\\\"board04149\\\",\\\"board02152\\\"]}\"]]","slots":[]},"result":"kp1545"}]}
{"type":"syscall","crankNum":5649,"vatID":"v25","deliveryNum":115,"syscallNum":0,"replay":false,"syscall":"send","ksc":["send","ko671",{"methargs":{"body":"#[\"setValue\",[\"{\\\"body\\\":\\\"#{\\\\\\\"liveOffers\\\\\\\":[],\\\\\\\"offerToPublicSubscriberPaths\\\\\\\":[],\\\\\\\"offerToUsedInvitation\\\\\\\":[],\\\\\\\"purses\\\\\\\":[{\\\\\\\"balance\\\\\\\":{\\\\\\\"brand\\\\\\\":\\\\\\\"$0.Alleged: Zoe Invitation brand\\\\\\\",\\\\\\\"value\\\\\\\":[{\\\\\\\"description\\\\\\\":\\\\\\\"oracle operator invitation\\\\\\\",\\\\\\\"installation\\\\\\\":\\\\\\\"$1.Alleged: BundleIDInstallation\\\\\\\",\\\\\\\"instance\\\\\\\":\\\\\\\"$2.Alleged: InstanceHandle\\\\\\\"}]},\\\\\\\"brand\\\\\\\":\\\\\\\"$0\\\\\\\"}]}\\\",\\\"slots\\\":[\\\"board0074\\\",\\\"board04149\\\",\\\"board02152\\\"]}\"]]","slots":[]},"result":"kp1546"}]}
```

No tests seem to cover the shape of data that the smart wallet publishes, at least not explicitly. Reading published invitation details is extensively tested indirectly.

### Upgrade Considerations
Requires an update of the smart wallet contract, which is also still needed for #12085 to become effective.

Should not affect any consumers.